requirements.txt
streamlit==1.41.1
google-ads==25.1.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
anthropic==0.42.0
pandas==2.2.3
numpy==2.2.1
plotly==5.24.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
reportlab==4.2.5
Pillow==11.1.0
